0.12.8 (#14)
* **prepare_cmake.sh** / **CMakeLists.txt** : + added '--no-shwild' / `CMAKE_NO_SHWILD` (#7)

* TODO : + more items (#8)


* + added skeleton for **UNIXem.Util** API (#9)


* Unit tests (#10)

* ~ renamed **test/component/test.component.glob/test.component.glob.cpp** => **test/component/test.component.glob/entry.cpp**

* ~ renamed **test/unit/*/*.cpp** => **test/component/*/entry.cpp**

* ~ renamed **test/component/test.component.dirent/entry.c** => **test/component/test.component.dirent/entry.cpp**

* **test.component.dirent** : + added two active test-cases

* **test.component.dirent** : + added more active test-cases

* **test.component.glob** : + added active test-case

* `glob()` : ~ trivial refactoring

* **test.component.glob** : + added more active test-case(s)

* **test.component.glob** : + added more active test-case(s)

---------


* Significant multiple improvements (#11)

* + added **UNIXem.Util.FS** implementation file to src/CMakeLists.txt

* **UNIXem.Util.FS** : + added `unixem_util_fs_char_is_path_sep()`

* ~ renamed TODO.txt => TODO.md

* TODO.md : + more items

* **test.component.unistd.mkdtemp** / **test.component.unistd.mkdtemp** : ~ minor mods

* + added **test.unit.util.fs**

* `glob()` : ~ simplifying refactoring

* `glob()` : ~ simplifying refactoring

* ~ preparatory refactoring

* + added **UNIXem.Util.Str** API

* + added `unixem_util_fs_get_home_directory()`

* `glob()` : ~ preparatory refactoring

* ~ fix

* `glob()` : ~ refactored in terms of `unixem_util_fs_get_home_directory()`

* + added and applied `unixem_util_fs_directory_exists()`

* ~ formatting

* version

* ~ MinGW compatibility

* ~ CMake dependencies fixes

---------


* Glob random (#13)

* + added **UNIXem.Util.FS** implementation file to src/CMakeLists.txt

* **UNIXem.Util.FS** : + added `unixem_util_fs_char_is_path_sep()`

* ~ renamed TODO.txt => TODO.md

* TODO.md : + more items

* **test.component.unistd.mkdtemp** / **test.component.unistd.mkdtemp** : ~ minor mods

* + added **test.unit.util.fs**

* `glob()` : ~ simplifying refactoring

* `glob()` : ~ simplifying refactoring

* ~ preparatory refactoring

* + added **UNIXem.Util.Str** API

* + added `unixem_util_fs_get_home_directory()`

* `glob()` : ~ preparatory refactoring

* ~ fix

* `glob()` : ~ refactored in terms of `unixem_util_fs_get_home_directory()`

* + added and applied `unixem_util_fs_directory_exists()`

* ~ formatting

* version

* `glob()` : ~ now implements `GLOB_NOSORT` by synthesising a random order of obtained results

* Merged 'dev' into 'glob-random'

Squashed commit of the following:

commit 6a89722cbb2d825fb7f61aee8441647c87a14d35
Author: Matt Wilson <152443343+mwsis@users.noreply.github.com>
Date:   Mon May 19 07:49:41 2025 +1000

    Significant multiple improvements (#11)

    * + added **UNIXem.Util.FS** implementation file to src/CMakeLists.txt

    * **UNIXem.Util.FS** : + added `unixem_util_fs_char_is_path_sep()`

    * ~ renamed TODO.txt => TODO.md

    * TODO.md : + more items

    * **test.component.unistd.mkdtemp** / **test.component.unistd.mkdtemp** : ~ minor mods

    * + added **test.unit.util.fs**

    * `glob()` : ~ simplifying refactoring

    * `glob()` : ~ simplifying refactoring

    * ~ preparatory refactoring

    * + added **UNIXem.Util.Str** API

    * + added `unixem_util_fs_get_home_directory()`

    * `glob()` : ~ preparatory refactoring

    * ~ fix

    * `glob()` : ~ refactored in terms of `unixem_util_fs_get_home_directory()`

    * + added and applied `unixem_util_fs_directory_exists()`

    * ~ formatting

    * version

    * ~ MinGW compatibility

    * ~ CMake dependencies fixes

    ---------


---------


* release

---------

+++ test/component/test.component.unistd.mkstemp/entry.c
@@ -1,14 +1,17 @@
 
 /*
- * Created:
- * Updated: 28th November 2024
+ * Created: ...
+ * Updated: 18th May 2025
  */
 
 #include <unistd.h>
 
-#include <xtests/xtests.h>
+#include <xtests/terse-api.h>
+
+#include <unixem/util/fs.h>
 
 #include <stdlib.h>
+
 
 int main(int argc, char** argv)
 {
@@ -29,7 +32,7 @@ int main(int argc, char** argv)
             {
                 int const e = errno;
 
-                fprintf(stderr, "mkstemp() failed: %s (%d)\n", strerror(e), e);
+                fprintf(stderr, "mkstemp() failed to create '%s': %s (%d)\n", template_path, strerror(e), e);
             }
             else
             {
@@ -51,29 +54,33 @@ int main(int argc, char** argv)
         /* Test-2 */
         if (XTESTS_CASE_BEGIN("Test-2", "testing mkstemp in specific directory"))
         {
-            char        template_path[] =   "C:\\temp\\abc.XXXX";
-            int const   r               =   mkstemp(template_path);
-
-            if (-1 == r)
+            if (unixem_util_fs_directory_exists("C:\\temp"))
             {
-                int const e = errno;
+                char        template_path[] =   "C:\\temp\\abc.XXXX";
+                int const   r               =   mkstemp(template_path);
 
-                fprintf(stderr, "mkstemp() failed: %s (%d)\n", strerror(e), e);
+                if (-1 == r)
+                {
+                    int const e = errno;
+
+                    fprintf(stderr, "mkstemp() failed to create '%s': %s (%d)\n", template_path, strerror(e), e);
+                }
+                else
+                {
+                    size_t const len = strlen(template_path);
+
+                    unixem_close(r);
+                    unlink(template_path);
+
+                    XTESTS_REQUIRE(XTESTS_TEST_INTEGER_EQUAL(16u, len));
+                    XTESTS_TEST_MULTIBYTE_STRING_EQUAL_N("C:\\temp\\abc.", template_path, 12u);
+                    XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 4]));
+                    XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 3]));
+                    XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 2]));
+                    XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 1]));
+                }
             }
-            else
-            {
-                size_t const len = strlen(template_path);
 
-                unixem_close(r);
-                unlink(template_path);
-
-                XTESTS_REQUIRE(XTESTS_TEST_INTEGER_EQUAL(16u, len));
-                XTESTS_TEST_MULTIBYTE_STRING_EQUAL_N("C:\\temp\\abc.", template_path, 12u);
-                XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 4]));
-                XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 3]));
-                XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 2]));
-                XTESTS_TEST_BOOLEAN_TRUE(isdigit(template_path[len - 1]));
-            }
 
             XTESTS_CASE_END("Test-2");
         }
